IGNITE-13549 CREATE INDEX command WIP
